ai-agent/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
openai==1.3.5
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
python-dateutil==2.8.2
# tavily-python==0.3.0  # Optional - requires Rust compiler. Install separately if needed.

